♻️ Move components to the components subdirectory
