+++ packages/react-resizin/src/components/index.ts
@@ -1,0 +1,3 @@
+export { default as ResizinContext } from './ResizinContext';
+export { default as ResizinProvider } from './Provider';
+export { default as Image } from './Image';
