+++ packages/react-resizin/src/index.ts
@@ -1,5 +1,2 @@
-export { default as ResizinContext } from './ResizinContext';
-export { default as ResizinProvider } from './Provider';
-export { default as Image } from './Image';
-
+export * from './components';
 export * from './hooks';
